+++ src/kreyolib/tokenize/__init__.py
@@ -1,3 +1,4 @@
 from kreyolib.tokenize.sentence import sent_tokenize
+from kreyolib.tokenize.word import word_tokenize
 
-__ALL__ = ["sent_tokenize"]
+__ALL__ = ["sent_tokenize", "word_tokenize"]